mjpeg-streaming/sdk/requirements.txt
Pillow
depthai-sdk==1.9.4.1